requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
python-jose>=3.3.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
stripe>=7.0.0
celery>=5.3.0
redis>=5.0.0
ffmpeg-python>=0.2.0
requests>=2.31.0
httpx>=0.26.0
